+++ src/main/groovy/ru/vyarus/gradle/plugin/animalsniffer/report/ReportCollector.groovy
@@ -22,6 +22,8 @@ class ReportCollector implements InvocationHandler {
 
     private static final int INFO = 2
 
+    // use to avoid registration of other threads output
+    String hostThread = Thread.currentThread().name
     // it should be org.gradle.api.internal.project.ant.AntLoggingAdapter
     Object originalListener
     // multiple signatures could be used for checking so exact signature is required
@@ -51,8 +53,8 @@ class ReportCollector implements InvocationHandler {
     Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
         Object event = args[0] // org.apache.tools.ant.BuildEvent
         // it should not be possible that other ant task will use this listener, but
-        // such case was reported (#3). Use ant task name to filter output
-        if (event.task?.taskName != 'animalsniffer') {
+        // such case was reported (#3). Use ant task name and thread name to filter output
+        if (event.task?.taskName != 'animalsniffer' || hostThread != Thread.currentThread().name) {
             if (originalListener) {
                 // redirect to original listener
                 method.invoke(originalListener, args)